pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hcai-lens"
authors = [
    {name = "Dominik Schiller", email = "dominik.schiller@uni-a.de"}
]
description = "LENS is a lightweight webserver designed to use Large Language Models as tool for data disocvery in human interactions."
readme = "README.md"
requires-python = ">=3.9"
keywords = ["LENS", "DISCOVER", "LLM", "machine learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3 :: Only"
]
dependencies = [
    "flask==3.0.0",
    "waitress",
    "python-dotenv",
    "psutil",
    "litellm==1.46.8"

]
dynamic = ["version"]

[tool.setuptools.packages.find]
#where = ["src"]  # list of folders that contain the packages (["."] by default)
#include = ["my_package*"]  # package names should match these glob patterns (["*"] by default)
exclude = ["*tests*", "local", "cache", "cml", "data", "log", "tmp"]  # exclude packages matching these glob patterns (empty by default)

[tool.setuptools.dynamic]
version = {attr = "lens.__version__"}

[project.urls]
#Documentation = "https://hcmlab.github.io/nova-server/docbuild/"
Repository = "https://github.com/hcmlab/lens"

[project.scripts]
lens = "nova_server.app:_run"
